- further eliminated occurences of AcidJunkie - Renamed playground projects to reflect what they are used for
