+++ .gitignore
@@ -1,4 +1,17 @@
 .vs
 .user
+*.user
 obj
 bin
+/shelf/
+workspace.xml
+# Rider ignored files
+/projectSettingsUpdater.xml
+/.idea.SyntaxInspectors.Analyzers.iml
+/contentModel.xml
+/modules.xml
+# Editor-based HTTP Client requests
+/httpRequests/
+# Datasource local storage ignored files
+/dataSources/
+/dataSources.local.xml
