--- src/.gitignore
@@ -1,14 +1,0 @@
-# Default ignored files
-/shelf/
-workspace.xml
-# Rider ignored files
-/projectSettingsUpdater.xml
-/.idea.AcidJunkie.Analyzers.iml
-/contentModel.xml
-/modules.xml
-# Editor-based HTTP Client requests
-/httpRequests/
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
-*.user
