--- src/.idea/.idea.AcidJunkie.Analyzers/.idea/.name
@@ -1,1 +1,0 @@
-AcidJunkie.Analyzers
