--- src/.idea/.idea.AcidJunkie.Analyzers/.idea/projectSettingsUpdater.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="RiderProjectSettingsUpdater">
-    <option name="singleClickDiffPreview" value="1" />
-    <option name="unhandledExceptionsIgnoreList" value="1" />
-    <option name="vcsConfiguration" value="3" />
-  </component>
-</project>
