+++ src/PlaygroundWithNugetPackage/.idea/.idea.Playground2.dir/.idea/.gitignore
