+++ src/PlaygroundWithNugetPackage/.idea/.idea.Playground2.dir/.idea/vcs.xml
