+++ src/PlaygroundWithProjectReference/PlaygroundWithProjectReference.csproj
